Clean up :)
Cleaned up double ups
