+++ MagicInventorySystem/OwnerMenu.cs
@@ -1,17 +1,22 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 
 namespace MagicInventorySystem
 {
-    internal class MainMenu
+    class OwnerMenu
     {
-        public void launchMenu()
+        public void Owner()
         {
-            Console.WriteLine("Welcome to the Magic Inventory System!");
+            Console.WriteLine("Welcome to the Magic Inventory System!(Owner)");
             Console.WriteLine("===========================================\n");
-            Console.WriteLine("     1. Owner\n");
-            Console.WriteLine("     2. Franchise Owner\n");
-            Console.WriteLine("     3. Customer\n");
-            Console.WriteLine("     4. Quit\n");
+            Console.WriteLine("     1. Display All Stock Requests\n");
+            Console.WriteLine("     2. Display Stock Requests (True / False)\n");
+            Console.WriteLine("     3. Display All Product Lines\n");
+            Console.WriteLine("     4. Return to Main Menu\n");
+            Console.WriteLine("     5. Exit\n");
 
             int programCounter = 0;
             while (programCounter == 0)
@@ -22,14 +27,12 @@ namespace MagicInventorySystem
                 {
                     userInput = Convert.ToInt32(Console.ReadLine());
                 }
-                catch(System.FormatException e)
+                catch (System.FormatException e)
                 {
                     Console.WriteLine("Invalid input entered. input type must be integer.\n");
                 }
                 if (userInput == 1)
                 {
-                    OwnerMenu owner = new OwnerMenu();
-                    owner.Owner();
                     programCounter++;
                 }
                 else if (userInput == 2)
@@ -44,14 +47,19 @@ namespace MagicInventorySystem
                 {
                     programCounter++;
                 }
-                else if(userInput > 4 || userInput < 1)
+
+                else if(userInput == 5)
+                {
+                    programCounter++;
+                }
+                else if (userInput > 5 || userInput < 1)
                 {
                     Console.WriteLine("Invalid input entered. please choose valid option.\n");
                 }
 
             }
             Console.ReadKey();
-
         }
+
     }
 }
