--- webDevTech-a1-dev/webDevTech-a1-dev/MagicInventorySystem.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MagicInventorySystem", "MagicInventorySystem\MagicInventorySystem.csproj", "{00295445-C82F-4E22-B828-6E7D8049CAA9}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{00295445-C82F-4E22-B828-6E7D8049CAA9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{00295445-C82F-4E22-B828-6E7D8049CAA9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{00295445-C82F-4E22-B828-6E7D8049CAA9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{00295445-C82F-4E22-B828-6E7D8049CAA9}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
