--- webDevTech-a1-dev/webDevTech-a1-dev/MagicInventorySystem/Driver.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace MagicInventorySystem
-{
-    class Driver
-    {
-        static void Main(string[] args)
-        {
-            MainMenu menu = new MainMenu();
-            menu.launchMenu();
-        }
-    }
-}
