--- webDevTech-a1-dev/webDevTech-a1-dev/MagicInventorySystem/bin/Debug/MagicInventorySystem.exe.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<configuration>
-    <startup> 
-        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.6.1" />
-    </startup>
-</configuration>
