--- webDevTech-a1-dev/webDevTech-a1-dev/README.md
@@ -1,2 +1,0 @@
-# webDevTech-a1
-Website Development Technologies Assignment 1
